+++ microservices/delivery-service/src/main/java/com/group40/deliveryservice/controller/UserController.java
@@ -1,17 +1,20 @@
 package com.group40.deliveryservice.controller;
 
+import com.group40.deliveryservice.dto.PersonResponse;
 import com.group40.deliveryservice.dto.UserRequest;
-import com.group40.deliveryservice.model.Customer;
-import com.group40.deliveryservice.model.Deliverer;
-import com.group40.deliveryservice.model.Dispatcher;
-import com.group40.deliveryservice.model.User;
+import com.group40.deliveryservice.model.*;
 import com.group40.deliveryservice.repository.UserRepository;
 import com.group40.deliveryservice.service.UserService;
 import lombok.RequiredArgsConstructor;
+import org.json.JSONException;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.io.IOException;
+import java.security.SecureRandom;
+import java.util.Base64;
 import java.util.List;
 
 //TODO: Change to ResponseEntity
@@ -25,40 +28,59 @@ public class UserController {
 
     private final UserRepository userRepository;
 
+    private final SecureRandom secureRandom;
+
+    private final Base64.Encoder base64Encoder;
+
     @GetMapping
     @ResponseStatus(HttpStatus.OK)
-    List<Customer> getAllCustomers() {
-        return userService.getAllCustomers();
+    ResponseEntity<?> getAllCustomers(@RequestHeader(HttpHeaders.AUTHORIZATION) String token) throws JSONException, IOException {
+        User user = userService.getUser(token);
+        if (user.getRole().equals(ERole.ROLE_DISPATCHER)) {
+            return ResponseEntity.ok(userService.getAllCustomers());
+        }else {
+            return ResponseEntity.badRequest().body("Not authorized!");
+        }
     }
 
     @DeleteMapping("/{id}")
-    void deleteUser(@PathVariable String id) {
-        userService.deleteUser(id);
+    ResponseEntity<?> deleteUser(@RequestHeader(HttpHeaders.AUTHORIZATION) String token, @PathVariable String id) throws JSONException, IOException {
+        User user = userService.getUser(token);
+        if (user.getRole().equals(ERole.ROLE_DISPATCHER)) {
+            userService.deleteUser(id);
+            return ResponseEntity.ok().build();
+        }else {
+            return ResponseEntity.badRequest().body("Not authorized!");
+        }
     }
 
     @PutMapping("/{id}")
-    User updateUser(@RequestBody UserRequest newUser, @PathVariable String id) {
+    ResponseEntity<?> updateUser(@RequestHeader(HttpHeaders.AUTHORIZATION) String token, @RequestBody UserRequest newUser, @PathVariable String id) throws JSONException, IOException {
+        User user = userService.getUser(token);
+        if (!user.getRole().equals(ERole.ROLE_DISPATCHER) && !user.getId().equals(id)) {
+            return ResponseEntity.badRequest().body("Not authorized!");
+        }
         switch (newUser.getRole()) {
             case ROLE_DELIVERER -> {
                 Deliverer deliverer = Deliverer.builder()
                         .email(newUser.getEmail())
                         .role(newUser.getRole())
                         .build();
-                return userService.updateUser(deliverer, id);
+                return ResponseEntity.ok(userService.updateUser(deliverer, id));
             }
             case ROLE_DISPATCHER -> {
                 Dispatcher dispatcher = Dispatcher.builder()
                         .email(newUser.getEmail())
                         .role(newUser.getRole())
                         .build();
-                return userService.updateUser(dispatcher, id);
+                return ResponseEntity.ok(userService.updateUser(dispatcher, id));
             }
             default -> {
                 Customer customer = Customer.builder()
                         .email(newUser.getEmail())
                         .role(newUser.getRole())
                         .build();
-                return userService.updateUser(customer, id);
+                return ResponseEntity.ok(userService.updateUser(customer, id));
             }
         }
     }
@@ -72,11 +94,14 @@ public class UserController {
         if(newUser.getPassword() == null || newUser.getPassword().isBlank() || newUser.getPassword().trim().isBlank()){
             throw new RuntimeException("Creating a new user requires a password.");
         }
+        byte[] randomBytes = new byte[64];
+        secureRandom.nextBytes(randomBytes);
         switch (newUser.getRole()) {
             case ROLE_DELIVERER -> {
                 Deliverer deliverer = Deliverer.builder()
                         .email(newUser.getEmail())
                         .role(newUser.getRole())
+                        .token(base64Encoder.encodeToString(randomBytes))
                         .build();
                 return userService.createUser(deliverer, newUser.getPassword());
             }
@@ -91,8 +116,26 @@ public class UserController {
                 Customer customer = Customer.builder()
                         .email(newUser.getEmail())
                         .role(newUser.getRole())
+                        .token(base64Encoder.encodeToString(randomBytes))
                         .build();
                 return userService.createUser(customer, newUser.getPassword());
+            }
+        }
+    }
+
+    @GetMapping
+    public ResponseEntity<?> getUserToken(@RequestHeader(HttpHeaders.AUTHORIZATION) String token) throws JSONException, IOException {
+        User user = userService.getUser(token);
+        switch (user.getRole()){
+            case ROLE_DELIVERER -> {
+                return ResponseEntity.ok(((Deliverer) user).getToken());
+            }
+            case ROLE_CUSTOMER -> {
+                return ResponseEntity.ok(((Customer) user).getToken());
+            }
+            default -> {
+                return ResponseEntity.badRequest()
+                        .body("Incorrect role");
             }
         }
     }
